requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.4
django-markdown-deux==1.0.6
django-markdownx==4.0.2
django-taggit==4.0.0
idna==3.4
Markdown==3.4.4
markdown2==2.4.10
markdownify==0.11.6
packaging==23.1
Pillow==10.0.0
plotly==5.16.1
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
tenacity==8.2.3
urllib3==2.0.4
